Deleted thermal.h, renamed fans.h, updated main.cpp
