--- Core/Inc/fan.h
@@ -1,8 +1,0 @@
-
-#ifndef HVC_FIRMWARE_2024_FAN_H
-#define HVC_FIRMWARE_2024_FAN_H
-
-/* Functions */
-void setFanRpm(float rpm);
-
-#endif //HVC_FIRMWARE_2024_FAN_H
