+++ Core/Inc/fans.h
@@ -1,0 +1,8 @@
+
+#ifndef HVC_FIRMWARE_2024_FANS_H
+#define HVC_FIRMWARE_2024_FANS_H
+
+/* Functions */
+void setFanRpm(float rpm);
+
+#endif //HVC_FIRMWARE_2024_FANS_H
