--- Core/Src/thermal.cpp
@@ -1,17 +1,0 @@
-
-#include "thermal.h"
-#include "cells.h"
-#include "fan.h"
-
-void updateParameters(float newMinTemp, float newMaxTemp) {
-    minAllowedTemp = newMinTemp;
-    maxAllowedTemp = newMaxTemp;
-}
-
-bool isTempWithinBounds() {
-    return getMinTemp() >= minAllowedTemp && getMaxTemp() <= maxAllowedTemp;
-}
-
-void thermalPeriodic() {
-    // TODO implement
-}
